editor.js, preparePage.js: no more eslint messages
